rename project from sky2cpp -> cpp2sky (#4)
