+++ .gitignore
@@ -37,7 +37,7 @@ user.bazelrc
 # bazel artifacts
 /bazel-bin
 /bazel-out
-/bazel-sky2cpp
+/bazel-cpp2sky
 /bazel-testlogs
 
 # editor settings
