+++ source/BUILD
@@ -4,6 +4,6 @@ cc_library(
   name = "tracer_lib",
   srcs = ["tracer_impl.cc"],
   hdrs = ["tracer_impl.h"],
-  deps = ["//sky2cpp:tracer_interface"],
+  deps = ["//cpp2sky:tracer_interface"],
   visibility = ["//visibility:public"],
 )
