+++ source/tracer_impl.h
@@ -1,4 +1,4 @@
-#include "sky2cpp/tracer.h"
+#include "cpp2sky/tracer.h"
 
 class TracerImpl : public Tracer {
  public:
